separate models into different files
