--- crategen/models.py
@@ -1,244 +1,0 @@
-from typing import Optional
-
-from pydantic import AnyUrl, BaseModel, Field, root_validator
-
-
-class WESRunLog(BaseModel):
-    """
-    Represents a run log in the Workflow Execution Service (WES).
-
-    Attributes:
-        name (Optional[str]): The name of the run.
-        start_time (Optional[str]): The start time of the run.
-        end_time (Optional[str]): The end time of the run.
-        cmd (Optional[list[str]]): The command executed in the run.
-        stdout (Optional[str]): The path to the stdout log.
-        stderr (Optional[str]): The path to the stderr log.
-        exit_code (Optional[int]): The exit code of the run.
-        tes_logs_url (Optional[str]): The URL of the TES logs.
-    """
-
-    name: Optional[str] = None
-    start_time: Optional[str] = None
-    end_time: Optional[str] = None
-    cmd: Optional[list[str]] = None
-    stdout: Optional[str] = None
-    stderr: Optional[str] = None
-    exit_code: Optional[int] = None
-    tes_logs_url: Optional[str] = None
-
-
-class WESOutputs(BaseModel):
-    """
-    Represents output files in WES.
-
-    Attributes:
-        location (str): The URL of the output file.
-        name (str): The name of the output file.
-    """
-
-    location: str
-    name: str
-
-
-class WESRequest(BaseModel):
-    """
-    Represents a workflow request in WES.
-
-    Attributes:
-        workflow_params (dict[str, str]): The parameters for the workflow.
-        workflow_type (str): The type of the workflow (e.g., CWL).
-        workflow_type_version (str): The version of the workflow type.
-        tags (Optional[dict[str, str]]): Additional tags associated with the workflow.
-    """
-
-    workflow_params: dict[str, str]
-    workflow_type: str
-    workflow_type_version: str
-    tags: Optional[dict[str, str]] = None
-
-
-class WESData(BaseModel):
-    """
-    Represents a WES run.
-
-    Attributes:
-        run_id (str): The unique identifier for the WES run.
-        request (WESRequest): The request associated with the WES run.
-        state (str): The state of the WES run.
-        run_log (WESRunLog): The log of the WES run.
-        task_logs (Optional[list[WESRunLog]]): The logs of individual tasks within the run.
-        outputs (list[WESOutputs]): The outputs of the WES run.
-    """
-
-    run_id: str
-    request: WESRequest
-    state: str
-    run_log: WESRunLog
-    task_logs: Optional[list[WESRunLog]] = Field(
-        None, description="This field is deprecated. Use tes_logs_url instead."
-    )
-    outputs: list[WESOutputs]
-
-    class Config:
-        extra = "allow"
-
-    @root_validator
-    def check_deprecated_fields(cls, values):
-        if values.get("task_logs") is not None:
-            print(
-                "DeprecationWarning: The 'task_logs' field is deprecated and will be removed in future versions. Use 'tes_logs_url' instead."
-            )
-        return values
-
-
-class WRROCInputs(BaseModel):
-    """
-    A model representing inputs in WRROC.
-
-    Attributes:
-        id (str): The unique identifier for the input.
-        name (str): The name of the input.
-    """
-
-    id: str
-    name: str
-
-
-class WRROCOutputs(BaseModel):
-    """
-    A model representing outputs in WRROC.
-
-    Attributes:
-        id (str): The unique identifier for the output.
-        name (str): The name of the output.
-    """
-
-    id: str
-    name: str
-
-
-class WRROCDataBase(BaseModel):
-    """
-    A base model representing common fields for WRROC entities.
-
-    Attributes:
-        id (str): The unique identifier for the WRROC entity.
-        name (str): The name of the WRROC entity.
-        description (Optional[str]): A brief description of the WRROC entity.
-        instrument (Optional[str]): The instrument used in the WRROC entity.
-        object (list[WRROCInputs]): A list of input objects related to the WRROC entity.
-        result (list[WRROCOutputs]): A list of output results related to the WRROC entity.
-        startTime (Optional[str]): The start time of the WRROC entity.
-        endTime (Optional[str]): The end time of the WRROC entity.
-        version (Optional[str]): The version of the WRROC entity.
-    """
-
-    id: str
-    name: str
-    description: Optional[str] = ""
-    instrument: Optional[str] = None
-    object: list[WRROCInputs]
-    result: list[WRROCOutputs]
-    startTime: Optional[str] = None
-    endTime: Optional[str] = None
-    version: Optional[str] = None
-
-    class Config:
-        extra = "allow"
-
-
-class WRROCData(WRROCDataBase):
-    """
-    A model representing a WRROC entity, inheriting from WRROCDataBase.
-    """
-
-    pass
-
-
-class WRROCDataTES(WRROCDataBase):
-    """
-    A model representing WRROC data specifically for TES conversion.
-
-    This model inherits from WRROCDataBase and includes all the necessary fields required for TES conversion.
-    """
-
-    pass
-
-
-class WRROCDataWES(WRROCDataBase):
-    """
-    A model representing WRROC data specifically for WES conversion.
-
-    This model inherits from WRROCDataBase and includes additional fields required for WES conversion.
-    """
-
-    status: str
-
-
-class WRROCProcess(BaseModel):
-    """
-    A model representing the WRROC Process Run profile.
-
-    Attributes:
-        id (str): The unique identifier for the WRROC entity.
-        name (str): The name of the WRROC entity.
-        description (Optional[str]): A brief description of the WRROC entity.
-        startTime (Optional[str]): The start time of the process.
-        endTime (Optional[str]): The end time of the process.
-        object (Optional[list[dict[str, str]]]): A list of input objects related to the process.
-        profiles (Optional[list[AnyUrl]]): URLs to the RO-Crate profiles used.
-    """
-
-    id: str
-    name: str
-    description: Optional[str] = ""
-    startTime: Optional[str] = None
-    endTime: Optional[str] = None
-    object: Optional[list[dict[str, str]]] = None
-    profiles: Optional[list[AnyUrl]] = None
-
-    class Config:
-        extra = "allow"
-
-
-class WRROCWorkflow(WRROCProcess):
-    """
-    A model representing the WRROC Workflow Run profile, inheriting from WRROCProcess.
-
-    Attributes:
-        workflowType (Optional[str]): The type of the workflow.
-        workflowVersion (Optional[str]): The version of the workflow.
-        result (Optional[list[dict[str, str]]]): A list of output results related to the workflow.
-        hasPart (Optional[list[AnyUrl]]): A list of parts or steps within the workflow.
-    """
-
-    workflowType: Optional[str] = None
-    workflowVersion: Optional[str] = None
-    result: Optional[list[dict[str, str]]] = None
-    hasPart: Optional[list[AnyUrl]] = None
-
-    class Config:
-        extra = "allow"
-
-
-class WRROCProvenance(WRROCWorkflow):
-    """
-    A model representing the WRROC Provenance Run profile, inheriting from WRROCWorkflow.
-
-    Attributes:
-        provenanceData (Optional[str]): Data related to the provenance of the workflow.
-        agents (Optional[list[dict[str, str]]]): A list of agents involved in the workflow.
-        activity (Optional[list[dict[str, str]]]): Activities related to the provenance.
-        generatedBy (Optional[list[AnyUrl]]): URLs of the entities that generated the data.
-        used (Optional[list[AnyUrl]]): URLs of the entities that were used in the data generation.
-    """
-
-    provenanceData: Optional[str] = None
-    agents: Optional[list[dict[str, str]]] = None
-    activity: Optional[list[dict[str, str]]] = None
-    generatedBy: Optional[list[AnyUrl]] = None
-    used: Optional[list[AnyUrl]] = None
-
-    class Config:
-        extra = "allow"
